Move out the nested class
